backend/requirements.txt
aniso8601==8.0.0
Click==7.0
DBUtils==1.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
